+++ lib/src/GenericDTO.dart
@@ -1,3 +1,0 @@
-class GenericDTO {
-  final Map<String, dynamic> values = {};
-}