+++ lib/transformers.dart
@@ -1,15 +1,18 @@
 import 'package:transformer_simplified/src/builtin_transformers.dart';
 
+/// Class to extend in order to define custom transformers
 abstract class Transformer<Input, Output> {
   Output? transform(Input? input);
 }
 
+/// Registry for the transformers.
+/// It contains a few builtin transformers
 class TransformerRegistry {
   static const String builtIn = "builtin://";
   static const String mapping = "mapping://";
   static const String custom = "custom://";
 
-  static TransformerRegistry registry = TransformerRegistry._();
+  static final TransformerRegistry _registry = TransformerRegistry._();
 
   final Map<String, Transformer> _builtinRegistry = {};
   final Map<String, Transformer> _customRegistry = {};
@@ -22,12 +25,16 @@ class TransformerRegistry {
     _builtinRegistry[BuiltInTransformer.toBool] = ToBool();
   }
 
+  static TransformerRegistry get() {
+    return _registry;
+  }
+
   Transformer? getBuiltIn(String key) {
     return _builtinRegistry[key];
   }
 
   void setCustom(String key, Transformer transformer) {
-    _builtinRegistry[key] = transformer;
+    _customRegistry[key] = transformer;
   }
 
   Transformer? getCustom(String key) {
@@ -35,6 +42,7 @@ class TransformerRegistry {
   }
 }
 
+/// List of the default builtin transformers
 class BuiltInTransformer {
   static const String toUppercase = 'toUppercase';
   static const String toLowercase = 'toLowercase';